\name{rvkstat_ks-package}
\alias{rvkstat_ks-package}
\alias{rvkstat_ks}
\docType{package}
\title{
\packageTitle{rvkstat}
}
\description{
\packageDescription{rvkstat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rvkstat}
\packageIndices{rvkstat}

}
\author{
\packageAuthor{rvkstat}

Maintainer: \packageMaintainer{rvkstat}
}

\keyword{ package }
