\name{rvkstat_ks-package}
\alias{rvkstat_ks-package}
\alias{rvkstat_ks}
\docType{package}
\title{
\packageTitle{rvkstat_ks}
}
\description{
\packageDescription{rvkstat_ks}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rvkstat_ks}
\packageIndices{rvkstat_ks}

}
\author{
\packageAuthor{rvkstat_ks}

Maintainer: \packageMaintainer{rvkstat_ks}
}

\keyword{ package }
